backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pydantic
pdfplumber
python-multipart
chromadb
fastembed
requests
